+++ backend/src/share/guard/shareOwner.guard.ts
@@ -1,5 +1,4 @@
 import {
-  CanActivate,
   ExecutionContext,
   Injectable,
   NotFoundException,
@@ -7,10 +6,17 @@ import {
 import { User } from "@prisma/client";
 import { Request } from "express";
 import { PrismaService } from "src/prisma/prisma.service";
+import { JwtGuard } from "../../auth/guard/jwt.guard";
+import { ConfigService } from "src/config/config.service";
 
 @Injectable()
-export class ShareOwnerGuard implements CanActivate {
-  constructor(private prisma: PrismaService) {}
+export class ShareOwnerGuard extends JwtGuard {
+  constructor(
+    configService: ConfigService,
+    private prisma: PrismaService,
+  ) {
+    super(configService);
+  }
 
   async canActivate(context: ExecutionContext) {
     const request: Request = context.switchToHttp().getRequest();
@@ -29,6 +35,8 @@ export class ShareOwnerGuard implements CanActivate {
     if (!share) throw new NotFoundException("Share not found");
 
     if (!share.creatorId) return true;
+
+    if (!(await super.canActivate(context))) return false;
 
     return share.creatorId == (request.user as User).id;
   }